[refactor] Matching Status View 전체 refactoring
