+++ Find-iOS/Find-iOS/Screen/MatchingStatus/Cell/Feeling/FeelingTVCell.xib
@@ -24,13 +24,13 @@
                 <autoresizingMask key="autoresizingMask"/>
                 <subviews>
                     <label opaque="NO" userInteractionEnabled="NO" contentMode="left" horizontalHuggingPriority="251" verticalHuggingPriority="251" text="받은 호감" textAlignment="natural" lineBreakMode="tailTruncation" baselineAdjustment="alignBaselines" adjustsFontSizeToFit="NO" translatesAutoresizingMaskIntoConstraints="NO" id="iOq-GI-fD0">
-                        <rect key="frame" x="16" y="20" width="67.5" height="21"/>
+                        <rect key="frame" x="16" y="20" width="70.5" height="23"/>
                         <fontDescription key="fontDescription" name="SpoqaHanSansNeo-Light" family="Spoqa Han Sans Neo" pointSize="18"/>
                         <color key="textColor" name="subGray5"/>
                         <nil key="highlightedColor"/>
                     </label>
                     <collectionView clipsSubviews="YES" multipleTouchEnabled="YES" contentMode="scaleToFill" showsHorizontalScrollIndicator="NO" dataMode="none" translatesAutoresizingMaskIntoConstraints="NO" id="DYt-jL-bUY">
-                        <rect key="frame" x="0.0" y="41" width="342" height="285"/>
+                        <rect key="frame" x="0.0" y="43" width="342" height="283"/>
                         <color key="backgroundColor" systemColor="systemBackgroundColor"/>
                         <collectionViewFlowLayout key="collectionViewLayout" scrollDirection="horizontal" minimumLineSpacing="10" minimumInteritemSpacing="10" id="9Fc-9I-Hx4">
                             <size key="itemSize" width="128" height="128"/>
