+++ Find-iOS/Find-iOS/Screen/MatchingStatus/ViewController/MatchingStatusVC.swift
@@ -7,13 +7,8 @@
 
 import UIKit
 
-enum MatchingCategory: Hashable {
-    case feelings, dibs
-}
-
 class MatchingStatusVC: UIViewController {
-    var toSendCategory: MatchingCategory?
-    
+    @IBOutlet weak var segueStackView: UIStackView!
     @IBOutlet var segIndicators: [UIView]!
     @IBOutlet var segueBtns: [UIButton]!
     @IBOutlet weak var matchingCV: UICollectionView! {
@@ -21,22 +16,23 @@ class MatchingStatusVC: UIViewController {
             matchingCV.delegate = self
             matchingCV.dataSource = self
             matchingCV.backgroundColor = .subGray6
+            matchingCV.register(MatchingDibsCVCell.nib(), forCellWithReuseIdentifier: MatchingDibsCVCell.identifier)
         }
     }
     
     override func viewDidLoad() {
         super.viewDidLoad()
         setSegueStyle()
-        whatSelected(seg: .feelings, idx: 0)
+        whatSelected(idx: 0)
         // Do any additional setup after loading the view.
     }
-
+    
     @IBAction func feelingsTapped(_ sender: Any) {
-        whatSelected(seg: .feelings, idx: 0)
+        whatSelected(idx: 0)
     }
     
     @IBAction func dibsTapped(_ sender: Any) {
-        whatSelected(seg: .dibs, idx: 1)
+        whatSelected(idx: 1)
     }
     
     // 페이징 관련 index 정의 함수
@@ -72,20 +68,17 @@ extension MatchingStatusVC {
     }
     
     // MARK: - Remind Selected Segue
-    func whatSelected(seg: MatchingCategory, idx: Int){
-        if toSendCategory != seg {
-            toSendCategory = seg
-            for i in 0..<2{
-                if i == idx{
-                    segIndicators[i].backgroundColor = .find_Purple
-                    segueBtns[i].setTitleColor(.find_Purple, for: .normal)
-                }else{
-                    segIndicators[i].backgroundColor = .subGray1
-                    segueBtns[i].setTitleColor(.subGray2, for: .normal)
-                }
+    func whatSelected(idx: Int){
+        for i in 0..<2{
+            if i == idx{
+                segIndicators[i].backgroundColor = .find_Purple
+                segueBtns[i].setTitleColor(.find_Purple, for: .normal)
+            }else{
+                segIndicators[i].backgroundColor = .subGray1
+                segueBtns[i].setTitleColor(.subGray2, for: .normal)
             }
-            NotificationCenter.default.post(name: NSNotification.Name("changeMatchingSegue"), object: toSendCategory)
         }
+        matchingCV.selectItem(at: IndexPath(item: idx, section: 0), animated: true, scrollPosition: .left)
     }
 }
 
@@ -95,15 +88,31 @@ extension MatchingStatusVC: UICollectionViewDelegate, UICollectionViewDataSource
     }
     
     func collectionView(_ collectionView: UICollectionView, cellForItemAt indexPath: IndexPath) -> UICollectionViewCell {
-        guard let cell = collectionView.dequeueReusableCell(withReuseIdentifier: MatchingStatusCVCell.identifier, for: indexPath) as? MatchingStatusCVCell else { return UICollectionViewCell() }
-        cell.curCategory = toSendCategory
-        return cell
+        if indexPath.row == 0 {
+            guard let cell = collectionView.dequeueReusableCell(withReuseIdentifier: MatchingStatusCVCell.identifier, for: indexPath) as? MatchingStatusCVCell else { return UICollectionViewCell() }
+            return cell
+        } else if indexPath.row == 1 {
+            guard let cell = collectionView.dequeueReusableCell(withReuseIdentifier: MatchingDibsCVCell.identifier, for: indexPath) as? MatchingDibsCVCell else { return UICollectionViewCell() }
+            return cell
+        }
+        return UICollectionViewCell()
     }
     
     // MARK: - Cell 크기
     func collectionView(_ collectionView: UICollectionView, layout collectionViewLayout: UICollectionViewLayout, sizeForItemAt indexPath: IndexPath) -> CGSize {
-        return CGSize(width: collectionView.frame.width, height: collectionView.frame.height)
+        return CGSize(width: self.view.frame.width, height: collectionView.frame.height)
     }
+    
+    //    // MARK: - Cell간의 상하간격 지정
+    //    func collectionView(_ collectionView: UICollectionView, layout collectionViewLayout: UICollectionViewLayout, minimumInteritemSpacingForSectionAt section: Int) -> CGFloat {
+    //        return 32
+    //    }
+    
+    //    // MARK: - 마진
+    //    func collectionView(_ collectionView: UICollectionView, layout collectionViewLayout: UICollectionViewLayout, insetForSectionAt section: Int) -> UIEdgeInsets
+    //    {
+    //        return UIEdgeInsets(top: -16, left: 0, bottom: 0, right: 0)
+    //    }
 }
 
 // MARK: - collectionView Horizontal Scrolling Magnetic Effect 적용
@@ -122,9 +131,9 @@ extension MatchingStatusVC: UIScrollViewDelegate{
             
             switch indexOfMajorCell {
             case 0:
-                whatSelected(seg: .feelings, idx: 0)
+                whatSelected(idx: 0)
             case 1:
-                whatSelected(seg: .dibs, idx: 1)
+                whatSelected(idx: 1)
             default:
                 return
             }
@@ -134,9 +143,9 @@ extension MatchingStatusVC: UIScrollViewDelegate{
             matchingCV.collectionViewLayout.collectionView!.scrollToItem(at: indexPath, at: .centeredHorizontally, animated: true)
             switch indexOfBeforeCell {
             case 0:
-                whatSelected(seg: .feelings, idx: 0)
+                whatSelected(idx: 0)
             case 1:
-                whatSelected(seg: .dibs, idx: 1)
+                whatSelected(idx: 1)
             default:
                 return
             }
